Crate stats

Structs

MStats
Geometric mean and standard deviations ratios

Med
Computed quartiles and median or time weighted versions of one column

Functions

amean
Arithmetic mean of an i64 slice

ameanstd

autocorr
Correlation coefficient of pairs of successive values of time series integer variable

awmean
Liearly weighted arithmetic mean of an i64 slice Linearly descending weights from the number of items down to one

awmeanstd

correlation
Correlation coefficient of a sample of two integer variables

gmean
Geometric mean of an i64 slice. The geometric mean is just an exponential of an arithmetic mean of log data (natural logarithms of the data items). The geometric mean is less sensitive to outliers near maximal value. However, we must normalise the data to avoid taking ln of zero, as minus infinity does not play nice on computers.

gmeanstd

gwmean
Time linearly weighted geometric mean of an i64 slice Linearly descending weights from the number of items down to one The geometric mean is just an exponential of an arithmetic mean of log data (natural logarithms of the data items) The geometric mean is less sensitive to outliers near maximal value

gwmeanstd

hmean
Harmonic mean of an i64 slice

hwmean
Linearly weighted harmonic mean of an f64 slice Linearly descending weights from the number of items down to one

median
Fast median (avoids sorting) Works on data discretised into ints [u32] with minimum and maximum [low, high]. high-low must be a moderate range, not exceeding u16size (= 65535)