Function stats::gmean

pub fn gmean(dvec: &[i64]) -> Result<f64>

Geometric mean of an i64 slice. The geometric mean is just an exponential of an arithmetic mean of log data (natural logarithms of the data items). The geometric mean is less sensitive to outliers near maximal value. However, we must normalise the data to avoid taking ln of zero, as minus infinity does not play nice on computers.