Struct stats::MinMax

pub struct MinMax<T> { /* fields omitted */ }

A commutative data structure for tracking minimum and maximum values.

This also stores the number of samples.

Methods

impl<T: PartialOrd + Clone> MinMax<T>

fn new() -> MinMax<T>

Create an empty state where min and max values do not exist.

fn add(&mut self, sample: T)

Add a sample to the data.

fn min(&self) -> Option<&T>

Returns the minimum of the data set.

None is returned if and only if the number of samples is 0.

fn max(&self) -> Option<&T>

Returns the maximum of the data set.

None is returned if and only if the number of samples is 0.

fn len(&self) -> usize

Returns the number of data point.