Struct stats::OnlineStats

pub struct OnlineStats { /* fields omitted */ }

Online state for computing mean, variance and standard deviation.

Methods

impl OnlineStats

fn new() -> OnlineStats

Create initial state.

Population size, variance and mean are set to 0.

fn from_slice<T: ToPrimitive>(samples: &[T]) -> OnlineStats

Initializes variance from a sample.

fn mean(&self) -> f64

Return the current mean.

fn stddev(&self) -> f64

Return the current standard deviation.

fn variance(&self) -> f64

Return the current variance.

fn add<T: ToPrimitive>(&mut self, sample: T)

Add a new sample.

fn add_null(&mut self)

Add a new NULL value to the population.

This increases the population size by 1.

fn len(&self) -> usize

Returns the number of data points.