Struct stats::Unsorted

pub struct Unsorted<T> { /* fields omitted */ }

A commutative data structure for lazily sorted sequences of data.

The sort does not occur until statistics need to be computed.

Note that this works on types that do not define a total ordering like f32 and f64. When an ordering is not defined, an arbitrary order is returned.

Methods

impl<T: PartialOrd> Unsorted<T>

fn new() -> Unsorted<T>

Create initial empty state.

fn add(&mut self, v: T)

Add a new element to the set.

fn len(&self) -> usize

Return the number of data points.

impl<T: PartialOrd + Eq + Clone> Unsorted<T>

fn cardinality(&mut self) -> usize

impl<T: PartialOrd + Clone> Unsorted<T>

fn mode(&mut self) -> Option<T>

Returns the mode of the data.

impl<T: PartialOrd + ToPrimitive> Unsorted<T>

fn median(&mut self) -> Option<f64>

Returns the median of the data.