Trait stats::Commute

pub trait Commute: Sized {
    fn merge(&mut self, other: Self);

    fn consume<I: Iterator<Item=Self>>(&mut self, other: I) { ... }
}

Defines an interface for types that have an identity and can be commuted.

The value returned by Default::default must be its identity with respect to the merge operation.

Required Methods

fn merge(&mut self, other: Self)

Merges the value other into self.

Provided Methods

fn consume<I: Iterator<Item=Self>>(&mut self, other: I)

Merges the values in the iterator into self.