Trait stats::Stats

pub trait Stats: IntoIterator + Clonewhere
    Self::Item: NumExt,{
    fn sum(&self) -> Self::Item { ... }
    fn count(&self) -> usize { ... }
    fn panicking_count(&self) -> Self::Item { ... }
    fn mean(&self) -> Self::Item { ... }
    fn checked_mean(&self) -> Option<Self::Item> { ... }
    fn variance(&self) -> Self::Item { ... }
    fn checked_variance(&self) -> Option<Self::Item> { ... }
}

A trait to be implemented for collection-like types that provides statistical methods. Requires that the type it is implemented on can be converted into an Iterator.

Provided Methods

fn sum(&self) -> Self::Item

Calculate the sum of all the items in the collection

Example

use stats::Stats;
assert_eq!(vec![1, 2, 3].sum(), 6);

use stats::Stats;
assert_eq!(vec![1.0, 2.0, 3.0].sum(), 6.0);

fn count(&self) -> usize

Count the items in the collection

Example

use stats::Stats;
assert_eq!(vec![1, 2, 3].count(), 3);

fn panicking_count(&self) -> Self::Item

Return the length of the collection as a Self::Item. Will panic if the length is greater than the maximum value of the type.

fn mean(&self) -> Self::Item

Find the mean of the collection

Examples

use stats::Stats;
assert_eq!(vec![1, 2, 3].mean(), 2);

use stats::Stats;
let v: Vec<f64> = vec![1.0, 2.0, 3.0];
assert_eq!(v.mean(), 2.0);

Watch out for integer division!

use stats::Stats;
assert_eq!(vec![1, 2, 3, 4].mean(), 2);

Panics

Panics if the collection is empty (has a length of 0).

use stats::Stats;
let v: Vec<i64> = vec![];
v.mean(); // panics

Will also panic if the length of the collection is too large to fit in Self::Item.

fn checked_mean(&self) -> Option<Self::Item>

Like Stats::mean, but returns None if the collection is empty, or if the length of the collection is too large to fit in Self::Item, whereas Stats::mean method will panic in these cases. The same caveat about integer division applies.

Examples

use stats::Stats;
assert_eq!(vec![1, 2, 3].checked_mean(), Some(2));

use stats::Stats;
assert_eq!(Vec::<i32>::new().checked_mean(), None);

use stats::Stats;
let v: Vec<f64> = vec![1.0, 2.0, 3.0];
assert_eq!(v.checked_mean(), Some(2.0));

fn variance(&self) -> Self::Item

Find the variance of the collection. The variance is the sum of the squared differences between each item and the mean, divided by the number of items in the collection. The variance is a measure of how spread out the items are.

Examples

use stats::Stats;
assert_eq!(vec![1.0, 2.0, 3.0].variance(), 2.0 / 3.0);

Panics

Panics if the collection is empty (has a length of 0). Will also panic if the length of the collection is too large to fit in Self::Item.

Wikipedia

fn checked_variance(&self) -> Option<Self::Item>

Like Stats::variance, but returns None if the collection is empty, or if the length of the collection is too large to fit in Self::Item, whereas Stats::variance method will panic in these cases.

Examples

use stats::Stats;
assert_eq!(vec![1.0, 2.0, 3.0].checked_variance(), Some(2.0 / 3.0));

use stats::Stats;
assert_eq!(Vec::<i32>::new().checked_variance(), None);